修改主键生成策略，增加update的样例

+++ asa-vwe-business-parent/asa-vwe-comments-parent/asa-vwe-comments-service/src/main/java/pers/demo/asa/vwe/comments/service/ICommentsService.java
@@ -30,5 +30,6 @@ public interface ICommentsService {
      * @param id id
      * @return 评论内容
      */
-    CommentsModel getById(long id);
+    CommentsModel getById(Integer id);
+
 }
